shield/backend/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
python-dotenv==1.0.0
psycopg2-binary==2.9.9
openai==0.28.1
pydantic==1.10.13
requests==2.31.0
python-jose==3.3.0
tiktoken==0.5.2
pytest==7.4.3
GitPython==3.1.40
gitdb==4.0.10
langchain==0.0.250
langchain-openai==0.0.2
autogen==0.2.0 